+++ Orleans.Streams.Kafka/Core/KafkaAdapter.cs
@@ -3,7 +3,9 @@ using Microsoft.Extensions.Logging;
 using Orleans.Serialization;
 using Orleans.Streams.Kafka.Config;
 using Orleans.Streams.Kafka.Producer;
+using Orleans.Streams.Kafka.Serialization;
 using Orleans.Streams.Utils;
+using Orleans.Streams.Utils.Serialization;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -18,6 +20,7 @@ namespace Orleans.Streams.Kafka.Core
 		private readonly SerializationManager _serializationManager;
 		private readonly ILoggerFactory _loggerFactory;
 		private readonly IGrainFactory _grainFactory;
+		private readonly IExternalStreamDeserializer _externalDeserializer;
 		private readonly IProducer<byte[], KafkaBatchContainer> _producer;
 		private readonly ILogger<KafkaAdapter> _logger;
 
@@ -31,7 +34,8 @@ namespace Orleans.Streams.Kafka.Core
 			IDictionary<string, QueueProperties> queueProperties,
 			SerializationManager serializationManager,
 			ILoggerFactory loggerFactory,
-			IGrainFactory grainFactory
+			IGrainFactory grainFactory,
+			IExternalStreamDeserializer externalDeserializer
 		)
 		{
 			_options = options;
@@ -39,6 +43,7 @@ namespace Orleans.Streams.Kafka.Core
 			_serializationManager = serializationManager;
 			_loggerFactory = loggerFactory;
 			_grainFactory = grainFactory;
+			_externalDeserializer = externalDeserializer;
 			_logger = _loggerFactory.CreateLogger<KafkaAdapter>();
 
 			Name = providerName;
@@ -66,8 +71,7 @@ namespace Orleans.Streams.Kafka.Core
 					streamGuid,
 					streamNamespace,
 					eventList,
-					requestContext,
-					false
+					requestContext
 				);
 
 				await _producer.Produce(batch);
@@ -90,7 +94,8 @@ namespace Orleans.Streams.Kafka.Core
 				_options,
 				_serializationManager,
 				_loggerFactory,
-				_grainFactory
+				_grainFactory,
+				_externalDeserializer
 			);
 
 		public void Dispose()