+++ Orleans.Streams.Kafka/Core/KafkaAdapterFactory.cs
@@ -3,9 +3,11 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Orleans.Configuration;
 using Orleans.Providers.Streams.Common;
+using Orleans.Runtime;
 using Orleans.Serialization;
 using Orleans.Streams.Kafka.Config;
 using Orleans.Streams.Utils;
+using Orleans.Streams.Utils.Serialization;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -20,6 +22,7 @@ namespace Orleans.Streams.Kafka.Core
 		private readonly SerializationManager _serializationManager;
 		private readonly ILoggerFactory _loggerFactory;
 		private readonly IGrainFactory _grainFactory;
+		private readonly IExternalStreamDeserializer _externalDeserializer;
 		private readonly IQueueAdapterCache _adapterCache;
 		private readonly IStreamQueueMapper _streamQueueMapper;
 		private readonly ILogger<KafkaAdapterFactory> _logger;
@@ -31,7 +34,8 @@ namespace Orleans.Streams.Kafka.Core
 			SimpleQueueCacheOptions cacheOptions,
 			SerializationManager serializationManager,
 			ILoggerFactory loggerFactory,
-			IGrainFactory grainFactory
+			IGrainFactory grainFactory,
+			IExternalStreamDeserializer externalDeserializer
 		)
 		{
 			_options = options ?? throw new ArgumentNullException(nameof(options));
@@ -40,6 +44,7 @@ namespace Orleans.Streams.Kafka.Core
 			_serializationManager = serializationManager;
 			_loggerFactory = loggerFactory;
 			_grainFactory = grainFactory;
+			_externalDeserializer = externalDeserializer;
 			_logger = loggerFactory.CreateLogger<KafkaAdapterFactory>();
 
 			if (options.Topics != null && options.Topics.Count == 0)
@@ -63,7 +68,8 @@ namespace Orleans.Streams.Kafka.Core
 				_queueProperties,
 				_serializationManager,
 				_loggerFactory,
-				_grainFactory
+				_grainFactory,
+				_externalDeserializer
 			);
 
 			return Task.FromResult<IQueueAdapter>(adapter);
@@ -82,12 +88,14 @@ namespace Orleans.Streams.Kafka.Core
 		{
 			var streamsConfig = services.GetOptionsByName<KafkaStreamOptions>(name);
 			var cacheOptions = services.GetOptionsByName<SimpleQueueCacheOptions>(name);
+			var deserializer = services.GetRequiredServiceByName<IExternalStreamDeserializer>(name);
 
 			var factory = ActivatorUtilities.CreateInstance<KafkaAdapterFactory>(
 				services,
 				name,
 				streamsConfig,
-				cacheOptions
+				cacheOptions,
+				deserializer
 			);
 
 			return factory;
@@ -105,7 +113,11 @@ namespace Orleans.Streams.Kafka.Core
 					var props = from kafkaTopic in meta.Topics
 								join userTopic in _options.Topics on kafkaTopic.Topic equals userTopic.Name
 								from partition in kafkaTopic.Partitions
-								select new QueueProperties(userTopic.Name, (uint)partition.PartitionId, userTopic.IsExternal);
+								select new QueueProperties(
+									userTopic.Name,
+									(uint)partition.PartitionId,
+									userTopic.IsExternal
+								);
 
 					return props.ToDictionary(prop => prop.QueueName);
 				}