+++ Orleans.Streams.Kafka/Serialization/SerializationContext.cs
@@ -1,0 +1,11 @@
+﻿using Orleans.Serialization;
+using Orleans.Streams.Utils.Serialization;
+
+namespace Orleans.Streams.Kafka.Serialization
+{
+	public struct SerializationContext
+	{
+		public SerializationManager SerializationManager { get; set; }
+		public IExternalStreamDeserializer ExternalStreamDeserializer { get; set; }
+	}
+}