+++ Samples/TestClient/Program.cs
@@ -71,8 +71,7 @@ namespace TestClient
 						{
 							options.BrokerList = new List<string> { "localhost:9092" };
 							options.ConsumerGroupId = "TestGroup";
-							options.ExternalMessageIdentifier = "external";
-							options.Topics = new List<Topic> { new Topic { Name = "gossip-testing" } };
+							options.Topics = new List<TopicConfig> { new TopicConfig { Name = "gossip-testing" } };
 						})
 						.Build();